+++ packages/react-three-rapier/src/components/Physics.tsx
@@ -20,7 +20,7 @@ import React, {
   useState
 } from "react";
 import { MathUtils, Matrix4, Object3D, Quaternion, Vector3 } from "three";
-import { useAsset } from "use-asset";
+import { suspend } from "suspend-react";
 import {
   CollisionPayload,
   CollisionEnterHandler,
@@ -31,7 +31,6 @@ import {
   RigidBodyAutoCollider,
   Vector3Tuple
 } from "../types";
-
 import {
   _matrix4,
   _position,
@@ -41,7 +40,8 @@ import {
 import {
   rapierQuaternionToQuaternion,
   useConst,
-  vectorArrayToVector3
+  vectorArrayToVector3,
+  vector3ToRapierVector
 } from "../utils/utils";
 import FrameStepper from "./FrameStepper";
 import { Debug } from "./Debug";
@@ -272,32 +272,37 @@ export interface PhysicsProps {
   gravity?: Vector3Tuple;
 
   /**
-   * The maximum velocity iterations the velocity-based constraint solver can make to attempt
-   * to remove the energy introduced by constraint stabilization.
-   *
-   * @defaultValue 1
+   * Amount of penetration the engine wont attempt to correct
+   * @defaultValue 0.001
    */
-  maxStabilizationIterations?: number;
+  allowedLinearError?: number;
 
   /**
-   * The maximum velocity iterations the velocity-based friction constraint solver can make.
-   *
-   * The greater this value is, the most realistic friction will be.
-   * However a greater number of iterations is more computationally intensive.
-   *
-   * @defaultValue 8
-   */
-  maxVelocityFrictionIterations?: number;
-
-  /**
-   * The maximum velocity iterations the velocity-based force constraint solver can make.
-   *
+   * The number of solver iterations run by the constraints solver for calculating forces.
    * The greater this value is, the most rigid and realistic the physics simulation will be.
    * However a greater number of iterations is more computationally intensive.
    *
    * @defaultValue 4
    */
-  maxVelocityIterations?: number;
+  numSolverIterations?: number;
+
+  /**
+   * Number of addition friction resolution iteration run during the last solver sub-step.
+   * The greater this value is, the most realistic friction will be.
+   * However a greater number of iterations is more computationally intensive.
+   *
+   * @defaultValue 4
+   */
+  numAdditionalFrictionIterations?: number;
+
+  /**
+   * Number of internal Project Gauss Seidel (PGS) iterations run at each solver iteration.
+   * Increasing this parameter will improve stability of the simulation. It will have a lesser effect than
+   * increasing `numSolverIterations` but is also less computationally expensive.
+   *
+   * @defaultValue 1
+   */
+  numInternalPgsIterations?: number;
 
   /**
    * The maximal distance separating two objects that will generate predictive contacts
@@ -308,10 +313,35 @@ export interface PhysicsProps {
   predictionDistance?: number;
 
   /**
-   * The Error Reduction Parameter in between 0 and 1, is the proportion of the positional error to be corrected at each time step
+   * Minimum number of dynamic bodies in each active island
+   *
+   * @defaultValue 128
+   */
+  minIslandSize?: number;
+
+  /**
+   * Maximum number of substeps performed by the solver
+   *
+   * @defaultValue 1
+   */
+  maxCcdSubsteps?: number;
+
+  /**
+   * The Error Reduction Parameter in between 0 and 1, is the proportion of the positional error to be corrected at each time step.
+   *
    * @defaultValue 0.8
    */
   erp?: number;
+
+  /**
+   * The approximate size of most dynamic objects in the scene.
+   *
+   * This value is used internally to estimate some length-based tolerance.
+   * This value can be understood as the number of units-per-meter in your physical world compared to a human-sized world in meter.
+   *
+   * @defaultValue 1
+   */
+  lengthUnit?: number;
 
   /**
    * Set the base automatic colliders for this physics world
@@ -395,13 +425,17 @@ export const Physics: FC<PhysicsProps> = (props) => {
     debug = false,
 
     gravity = [0, -9.81, 0],
-    maxStabilizationIterations = 1,
-    maxVelocityFrictionIterations = 8,
-    maxVelocityIterations = 4,
+    allowedLinearError = 0.001,
     predictionDistance = 0.002,
-    erp = 0.8
+    numSolverIterations = 4,
+    numAdditionalFrictionIterations = 4,
+    numInternalPgsIterations = 1,
+    minIslandSize = 128,
+    maxCcdSubsteps = 1,
+    erp = 0.8,
+    lengthUnit = 1
   } = props;
-  const rapier = useAsset(importRapier);
+  const rapier = suspend(importRapier, ["@react-thee/rapier", importRapier]);
   const { invalidate } = useThree();
 
   const rigidBodyStates = useConst<RigidBodyStateMap>(() => new Map());
@@ -463,23 +497,34 @@ export const Physics: FC<PhysicsProps> = (props) => {
 
   // Update mutable props
   useEffect(() => {
-    worldProxy.gravity = vectorArrayToVector3(gravity);
-    worldProxy.integrationParameters.maxStabilizationIterations =
-      maxStabilizationIterations;
-    worldProxy.integrationParameters.maxVelocityFrictionIterations =
-      maxVelocityFrictionIterations;
-    worldProxy.integrationParameters.maxVelocityIterations =
-      maxVelocityIterations;
-    worldProxy.integrationParameters.predictionDistance = predictionDistance;
+    worldProxy.gravity = vector3ToRapierVector(gravity);
+
+    worldProxy.integrationParameters.numSolverIterations = numSolverIterations;
+    worldProxy.integrationParameters.numAdditionalFrictionIterations =
+      numAdditionalFrictionIterations;
+    worldProxy.integrationParameters.numInternalPgsIterations =
+      numInternalPgsIterations;
+
+    worldProxy.integrationParameters.normalizedAllowedLinearError =
+      allowedLinearError;
+    worldProxy.integrationParameters.minIslandSize = minIslandSize;
+    worldProxy.integrationParameters.maxCcdSubsteps = maxCcdSubsteps;
+    worldProxy.integrationParameters.normalizedPredictionDistance =
+      predictionDistance;
     worldProxy.integrationParameters.erp = erp;
+    worldProxy.lengthUnit = lengthUnit;
   }, [
     worldProxy,
     ...gravity,
-    maxStabilizationIterations,
-    maxVelocityIterations,
-    maxVelocityFrictionIterations,
+    numSolverIterations,
+    numAdditionalFrictionIterations,
+    numInternalPgsIterations,
+    allowedLinearError,
+    minIslandSize,
+    maxCcdSubsteps,
     predictionDistance,
-    erp
+    erp,
+    lengthUnit
   ]);
 
   const getSourceFromColliderHandle = useCallback((handle: ColliderHandle) => {