+++ packages/react-three-rapier/src/utils/utils-collider.ts
@@ -138,6 +138,12 @@ const mutableColliderOptions: MutableColliderOptions = {
   restitutionCombineRule: (collider, value) => {
     collider.setRestitutionCombineRule(value);
   },
+  activeCollisionTypes: (collider, value: number) => {
+    collider.setActiveCollisionTypes(value);
+  },
+  contactSkin: (collider, value: number) => {
+    collider.setContactSkin(value);
+  },
   // To make sure the options all mutable options are listed
   quaternion: () => {},
   position: () => {},
@@ -476,6 +482,7 @@ export const cleanRigidBodyPropsForCollider = (props: RigidBodyProps = {}) => {
     canSleep,
     ccd,
     gravityScale,
+    softCcdPrediction,
     ...rest
   } = props;