+++ packages/react-three-rapier-addons/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @react-three/rapier-addons
 
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [bad7da7]
+  - @react-three/rapier@1.3.1
+
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [1517acb]
+  - @react-three/rapier@1.3.0
+
+## 4.0.0-canary.0
+
+### Patch Changes
+
+- Updated dependencies [1517acb]
+  - @react-three/rapier@1.3.0-canary.0
+
 ## 3.0.3
 
 ### Patch Changes